// actyondatafront/src/app/shared/models/cliente/Dados-Profissionais-Model.ts
export class DadosProissionaisModel {
    EMPRESA: string;
    CARGO: string;
    VALOR_RENDA: number;
    CEP_EMPRESA: string;
    ENDERECO_EMPRESA: string;
    NUMERO_EMPRESA: string;
    COMPLEMENTO_ENDERECO_EMPRESA: string;
    BAIRRO_EMPRESA: string;
    CIDADE_EMPRESA: string;
    UF_EMPRESA: string;
}